Added HTTPS support to OpenGTS sender
Consolidated OpenGTS to use CustomURL sender for HTTP and HTTPS

Issue #407
