+++ gpslogger/src/main/java/com/mendhak/gpslogger/loggers/customurl/CustomUrlJob.java
@@ -1,7 +1,6 @@
 package com.mendhak.gpslogger.loggers.customurl;
 
 
-import com.mendhak.gpslogger.common.SerializableLocation;
 import com.mendhak.gpslogger.common.Strings;
 import com.mendhak.gpslogger.common.events.UploadEvents;
 import com.mendhak.gpslogger.common.slf4j.Logs;
@@ -19,17 +18,18 @@ import java.io.IOException;
 public class CustomUrlJob extends Job {
 
     private static final Logger LOG = Logs.of(CustomUrlJob.class);
-    private SerializableLocation loc;
     private String logUrl;
     private String basicAuthUser;
     private String basicAuthPassword;
+    private UploadEvents.BaseUploadEvent callbackEvent;
 
 
-    public CustomUrlJob(String logUrl, String basicAuthUser, String basicAuthPassword) {
+    public CustomUrlJob(String logUrl, String basicAuthUser, String basicAuthPassword, UploadEvents.BaseUploadEvent callbackEvent ) {
         super(new Params(1).requireNetwork().persist());
         this.logUrl = logUrl;
         this.basicAuthPassword = basicAuthPassword;
         this.basicAuthUser = basicAuthUser;
+        this.callbackEvent = callbackEvent;
     }
 
     @Override
@@ -70,13 +70,14 @@ public class CustomUrlJob extends Job {
 
         if (response.isSuccessful()) {
             LOG.debug("Success - response code " + response);
-            EventBus.getDefault().post(new UploadEvents.CustomUrl().succeeded());
+            EventBus.getDefault().post(callbackEvent.succeeded());
         }
         else {
             LOG.error("Unexpected response code " + response );
-            EventBus.getDefault().post(new UploadEvents.CustomUrl().failed("Unexpected code " + response,new Throwable(response.body().string())));
-
+            EventBus.getDefault().post(callbackEvent.failed("Unexpected code " + response,new Throwable(response.body().string())));
         }
+
+        response.body().close();
 
     }
 
@@ -87,7 +88,7 @@ public class CustomUrlJob extends Job {
 
     @Override
     protected boolean shouldReRunOnThrowable(Throwable throwable) {
-        EventBus.getDefault().post(new UploadEvents.CustomUrl().failed("Could not send to custom URL", throwable));
+        EventBus.getDefault().post(callbackEvent.failed("Could not send to custom URL", throwable));
         LOG.error("Could not send to custom URL", throwable);
         return true;
     }
