+++ gpslogger/src/main/res/xml/opengtssettings.xml
@@ -22,7 +22,7 @@
         <com.afollestad.materialdialogs.prefs.MaterialListPreference android:key="opengts_server_communication_method"
                         android:title="@string/opengts_server_communication_method"
                         android:summary="@string/opengts_server_communication_method_summary"
-                        android:entries="@array/opengts_server_communication_methods" android:defaultValue="0"
+                        android:entries="@array/opengts_server_communication_methods_values" android:defaultValue="0"
                         android:entryValues="@array/opengts_server_communication_methods_values"/>
         <com.afollestad.materialdialogs.prefs.MaterialEditTextPreference android:key="autoopengts_server_path"
                             android:title="@string/autoopengts_server_path"
